refactor: cleanup logic
Why:
- repetitive work in defining variables
- repetitive work in confirming template is correct (repeated create/delete steps)
- need to find ways to pass environment variables to template

How:
- add computed variables
- add options to perform dry run, which will print to the console
- provide option to read from environment variables

+++ src/gen.js
@@ -4,7 +4,7 @@ import inquirer from "inquirer";
 import path from "path";
 
 import handlebars from "./template.js";
-import { open, load, ext } from "./file";
+import { createFolder, folderName, open, load, ext } from "./file";
 import { mergeObject, isObject, toObject } from "./object.js";
 
 const UTF8 = { encoding: "utf8" };
@@ -14,7 +14,8 @@ const FileStatus = {
   EXISTS: chalk.red("(exists)")
 };
 
-export const gen = async (configPath = "config.json", commandType, name) => {
+export const gen = async (options, type, name) => {
+  const { config: configPath = "config.json", dryRun } = options;
   const config = load(configPath);
 
   // Allow extending of handlebars if a config.js file is loaded.
@@ -28,25 +29,29 @@ export const gen = async (configPath = "config.json", commandType, name) => {
     default:
   }
 
-  const command = config[commandType];
+  const command = config[type];
   if (!command)
-    throw new Error(
-      `Command "${commandType} is not specified in the ${configPath}"`
-    );
+    throw new Error(`Command "${type} is not specified in the ${configPath}"`);
 
-  const { prompts, actions } = command;
+  const { prompts, actions, environment } = command;
 
   const answers = prompts ? await inquirer.prompt(prompts) : {};
+  const envvars = parseEnvironment(environment);
 
   const templateStatus = {};
   const destinationStatus = {};
 
   for (const action of actions) {
-    const data = {
+    const initialData = {
       name,
-      command: commandType,
+      type,
       ...answers,
-      ...action.data
+      ...envvars
+    };
+    const variables = parseVariables(handlebars, initialData, action.variables);
+    const data = {
+      ...initialData,
+      ...variables
     };
 
     let template;
@@ -58,6 +63,7 @@ export const gen = async (configPath = "config.json", commandType, name) => {
     const destinationPath = handlebars.compile(action.destination)(data);
 
     try {
+      await createFolder(folderName(templatePath));
       template = await open(templatePath, "a+");
       content = await template.readFile(UTF8);
       templateStatus[templatePath] = content.length
@@ -70,18 +76,27 @@ export const gen = async (configPath = "config.json", commandType, name) => {
     }
 
     if (!content.trim().length) {
+      console.log(chalk.red(`Skipping template: ${templatePath} is empty`));
       continue;
     }
 
-    try {
-      destination = await open(destinationPath, "wx+");
-      await destination.writeFile(handlebars.compile(content)(data), UTF8);
+    if (dryRun) {
+      console.log(chalk.blue(`Printing ${templatePath}:`));
+      console.log();
+      console.log(handlebars.compile(content)(data));
+      console.log();
+    } else {
+      try {
+        await createFolder(folderName(destinationPath));
+        destination = await open(destinationPath, "wx+");
+        await destination.writeFile(handlebars.compile(content)(data), UTF8);
 
-      destinationStatus[destinationPath] = FileStatus.CREATED;
-    } catch (error) {
-      destinationStatus[destinationPath] = FileStatus.EXISTS;
-    } finally {
-      destination?.close();
+        destinationStatus[destinationPath] = FileStatus.CREATED;
+      } catch (error) {
+        destinationStatus[destinationPath] = FileStatus.EXISTS;
+      } finally {
+        destination?.close();
+      }
     }
   }
 
@@ -101,4 +116,32 @@ function printTree(obj = {}, msg = "") {
   }
   console.log(chalk.blue(msg));
   console.log(treeify.asTree(tree, true));
+}
+
+export async function createConfig() {
+  const config = require("../config.sample.json5");
+  // Opens the file for read/write, but throws error when exists.
+  const file = await open("config.json", "wx+");
+  await file.writeFile(JSON.stringify(config, null, 2), UTF8);
+}
+
+function parseEnvironment(obj = {}) {
+  const entries = Object.entries(obj).map(([key, value]) => {
+    const environmentValue = process.env[value];
+    if (!environmentValue) {
+      throw new Error(`Env "${value}" is defined, but no value is provided`);
+    }
+    return [key, environmentValue];
+  });
+  return Object.fromEntries(entries);
+}
+
+// parseVariables parses the custom template variables that are action
+// specific.
+function parseVariables(handlebars, data, obj = {}) {
+  const entries = Object.entries(obj).map(([key, value]) => {
+    const parsed = handlebars.compile(value)(data);
+    return [key, parsed];
+  });
+  return Object.fromEntries(entries);
 }
